Refactor to prevent unnecessary renders
